Moved model objects into model folder
