+++ src/commands/showHierarchyBase.ts
@@ -1,116 +1,11 @@
 import { CommandBase } from '@commands';
 import { Config, FileSystemUtils } from '@src';
+import { Edge, Node } from '@model';
 import { Base64 } from 'js-base64';
 import * as fs from 'fs';
 import * as path from 'path';
 import * as vscode from 'vscode';
-import { Project } from '@model';
 
-export enum NodeType {
-  none,
-  component,
-  directive,
-  injectable,
-  module,
-  pipe,
-  rootNode
-}
-export class Node {
-  private config: Config = new Config();
-  constructor(id: string, name: string, isRoot: boolean, nodeType: NodeType = NodeType.none) {
-    this.id = id;
-    this.name = name;
-    this.isRoot = isRoot;
-    this.nodeType = nodeType;
-  }
-  public id: string;
-  public name: string;
-  public isRoot: boolean;
-  public nodeType: NodeType;
-
-  public toJsonString(): string {
-    let nodeColorAttr = '';
-    switch (this.nodeType) {
-      case NodeType.rootNode:
-        nodeColorAttr = `, color: "${this.config.rootNodeBackgroundColor}", shape: "${this.config.visNodeShape}"`;
-        break;
-      case NodeType.component:
-        nodeColorAttr = `, color: "${this.config.componentNodeBackgroundColor}", shape: "${this.config.componentNodeShape}"`;
-        break;
-      case NodeType.module:
-        nodeColorAttr = `, color: "${this.config.moduleNodeBackgroundColor}", shape: "${this.config.moduleNodeShape}"`;
-        break;
-      case NodeType.pipe:
-        nodeColorAttr = `, color: "${this.config.pipeNodeBackgroundColor}", shape: "${this.config.pipeNodeShape}"`;
-        break;
-      case NodeType.directive:
-        nodeColorAttr = `, color: "${this.config.directiveNodeBackgroundColor}", shape: "${this.config.directiveNodeShape}"`;
-        break;
-      default:
-        nodeColorAttr = '';
-        break;
-    }
-    const label = this.config.maximumNodeLabelLength !== -1 && this.name.length > this.config.maximumNodeLabelLength ? this.name.substr(0, this.config.maximumNodeLabelLength) + '...' : this.name;
-    return `{id: "${this.id}", label: "${label}" ${nodeColorAttr}}`;
-  }
-
-  public static getNodeType(project: Project, className: string) {
-    let nodeType = NodeType.none;
-    if (project.moduleNames.has(className) || className.endsWith('Module') || className.includes("Module.")) {
-      nodeType = NodeType.module;
-    }
-    else if (project.components.has(className) || className.endsWith('Component')) {
-      nodeType = NodeType.component;
-    }
-    else if (project.directives.has(className) || className.endsWith('Directive')) {
-      nodeType = NodeType.directive;
-    }
-    else if (project.pipes.has(className) || className.endsWith('Pipe')) {
-      nodeType = NodeType.pipe;
-    }
-    else if (project.injectables.has(className)) {
-      nodeType = NodeType.injectable;
-    }
-    return nodeType;
-  }
-}
-
-export enum ArrowType {
-  none = 0,
-  import = 1,
-  export = 2,
-  injectable = 3
-}
-
-export class Edge {
-  private config: Config = new Config();
-  constructor(id: string, source: string, target: string, arrowType: ArrowType = ArrowType.none) {
-    this.id = id;
-    this.source = source;
-    this.target = target;
-    this.arrowType = arrowType;
-  }
-  public id: string;
-  public source: string;
-  public target: string;
-  public arrowType: ArrowType;
-
-  public toJsonString(): string {
-    let arrowColorAttr = '';
-    switch (this.arrowType) {
-      case ArrowType.import:
-        arrowColorAttr = `, color: "${this.config.importEdgeColor}"`;
-        break;
-      case ArrowType.export:
-        arrowColorAttr = `, color: "${this.config.exportEdgeColor}"`;
-        break;
-      default:
-        arrowColorAttr = '';
-        break;
-    }
-    return `{from: "${this.source}", to: "${this.target}", arrows: arrowAttr ${arrowColorAttr} }`;
-  }
-}
 
 export class ShowHierarchyBase extends CommandBase {
   protected fsUtils: FileSystemUtils = new FileSystemUtils();
@@ -166,7 +61,7 @@ export class ShowHierarchyBase extends CommandBase {
   }
 
   protected saveAsDgml(dgmlGraphFilename: string, messageText: string) {
-    
+
   }
   
   protected generateHtmlContent(webview: vscode.Webview, outputJsFilename: string): string {
