+++ src/model/ArrowType.ts
@@ -1,0 +1,6 @@
+export enum ArrowType {
+  none = 0,
+  import = 1,
+  export = 2,
+  injectable = 3
+}
