+++ src/model/NodeType.ts
@@ -1,0 +1,9 @@
+export enum NodeType {
+  none,
+  component,
+  directive,
+  injectable,
+  module,
+  pipe,
+  rootNode
+}
