+++ src/model/index.ts
@@ -1,3 +1,7 @@
-export * from './NamedEntity';
+export * from './ArrowType';
 export * from './Component';
+export * from './Edge';
+export * from './NamedEntity';
+export * from './Node';
+export * from './NodeType';
 export * from './Project';
